Working first version (w/ button "motion")
This version works in its entirety (the timing is a little off of the goal but it's close enough)
The P2.3 simulates "motion," so it doesn't work with an infrared sensor as of right now. Every time you press the button it detects "motion," so spamming the button for ~10 seconds will put it in the alert state.